Moving the definition of TemplateExpressionError to avoid circular reference
