+++ pyomo/core/expr/numvalue.py
@@ -25,6 +25,13 @@ from pyomo.core.expr.expr_common import \
 
 logger = logging.getLogger('pyomo.core')
 
+
+class TemplateExpressionError(ValueError):
+
+    def __init__(self, template, *args, **kwds):
+        self.template = template
+        super(TemplateExpressionError, self).__init__(*args, **kwds)
+
 def _generate_sum_expression(etype, _self, _other):
     raise RuntimeError("incomplete import of Pyomo expression system")  #pragma: no cover
 def _generate_mul_expression(etype, _self, _other):
@@ -217,10 +224,6 @@ def value(obj, exception=True):
         #
         # Here, we try to catch the exception
         #
-
-        # This import needs to be deferred to here due to circular
-        # imports
-        from pyomo.core.expr.current import TemplateExpressionError
 
         try:
             tmp = obj(exception=True)
